+++ app/service/test/Test.php
@@ -1,0 +1,18 @@
+<?php
+namespace My\Service\Test;
+
+#use Flex\Banana\Classes\Json\JsonEncoder;
+
+class Test {
+    public function __construct() {
+
+    }
+
+    public function do() : string {
+
+        return json_encode([
+            "result" => "true",
+            "msg" => "Test"
+        ]);
+    }
+}