+++ app/env/env.php
@@ -1,7 +1,0 @@
-<?php
-# define
-define('DB_HOSTNAME','flexreact-php-mysql:test_db');
-define('DB_USERID','test');
-define('DB_PASSWORD','test!@!@');
-define('DB_PORT',3306);
-?>