+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### - 2024-10-18
 - env/env.php composer.json autoload 등록 및 테스트
+- 데이터 베이스 연결 관련 리스트,수정,업데이트,저장,삭제 테스트 test/Db class new add
 
 ### - 2024-10-17
 - 내 네임스페이스와 경로 composer.json 설정 및 라우터 실행 관련 업데이트